requirements.txt
playwright==1.49.0
playwright-stealth==2.0.3
schedule==1.2.2
pyyaml==6.0.2
python-dotenv==1.0.0
matplotlib==3.9.0
japanize-matplotlib==1.1.3
flask==3.1.0

